Adding benchmarking to hipsparse Part 4 (#503)
* Add display info to all testing_xxx files

* Clang formatting

* Add hipsparse bench app to allow writing benchmarking results to json files for plotting

* Clang formatting

* Add missing code to client.cpp file to allow use of --bench-x

* Clang formatting

* Remove unneeded define in hipsparse.h

* Clang formatting

* Remove commented header include

---------


[ROCm/hipSPARSE commit: a7b00f0cbc70e805bd2e873e9686c5b6464bcf9e]

+++ clients/benchmarks/hipsparse_bench_cmdlines.cpp
@@ -15,7 +15,7 @@ int hipsparse_bench_cmdlines::get_nsamples() const
 {
     return this->m_cmd.get_nsamples();
 };
-int hipsparse_bench_cmdlines::get_option_index_x() const
+size_t hipsparse_bench_cmdlines::get_option_index_x() const
 {
     return this->m_cmd.get_option_index_x();
 };
